# Small optimization